requirements.txt
absl-py==2.1.0
alembic==1.14.1
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
filelock==3.17.0
filterpy==1.4.5
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.2.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
h5py==3.13.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jax==0.5.0
jax-cuda12-pjrt==0.5.0
jax-cuda12-plugin==0.5.0
jaxlib==0.5.0
Jinja2==3.1.5
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
lxml==5.3.1
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
moviepy==2.1.2
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy~=1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.8.61
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
optuna==4.2.1
packaging==24.2
pandas==2.2.3
pillow==10.4.0
proglog==0.1.10
protobuf==5.29.3
psutil==7.0.0
py-cpuinfo==9.0.0
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
SQLAlchemy==2.0.38
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-hub==0.16.1
termcolor==2.5.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tifffile==2025.2.18
torch~=2.2.2
torchvision~=0.17.2
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.80
ultralytics-thop==2.0.14
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
boxmot~=12.0.1